create unified compile options
