+++ toolchain/driver/compile_options.h
@@ -2,8 +2,8 @@
 // Exceptions. See /LICENSE for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef CARBON_TOOLCHAIN_DRIVER_SHARED_COMPILE_OPTIONS_H_
-#define CARBON_TOOLCHAIN_DRIVER_SHARED_COMPILE_OPTIONS_H_
+#ifndef CARBON_TOOLCHAIN_DRIVER_COMPILE_OPTIONS_H_
+#define CARBON_TOOLCHAIN_DRIVER_COMPILE_OPTIONS_H_
 
 #include <memory>
 
@@ -26,13 +26,25 @@ namespace Carbon {
 // `build` and `compile` subcommands, supporting different flags
 // for each subcomand.
 //
-// Note that the `Build()` function provides only the common flags shared
-// between both subcommands, and each subcommand provides their own specialized
-// flags in their respective `Options` structs.
+// Note that the each subcommand supports its own `Build*()` function,
+// supporting the different use cases for command-line control of compilation.
 //
-// Members are documented in the `Build` function.
-struct SharedCompileOptions {
-  auto Build(CommandLine::CommandBuilder& b) -> void;
+// Members are documented in their respective `Build` functions.
+struct CompileOptions {
+  enum class Phase : int8_t {
+    Lex,
+    Parse,
+    Check,
+    Lower,
+    Optimize,
+    CodeGen,
+  };
+
+  friend auto operator<<(llvm::raw_ostream& out, Phase phase)
+      -> llvm::raw_ostream&;
+
+  auto BuildForCompileSubcommand(CommandLine::CommandBuilder& b) -> void;
+  auto BuildForBuildSubcommand(CommandLine::CommandBuilder& b) -> void;
 
   // Validate the target before passing to clang.
   auto ValidateTarget(Diagnostics::NoLocEmitter& emitter)
@@ -58,14 +70,50 @@ struct SharedCompileOptions {
   bool include_debug_info = true;
   bool run_llvm_verifier = true;
 
+  Phase phase = Phase::CodeGen;
+  Check::CheckParseTreesOptions::DumpSemIRRanges dump_sem_ir_ranges;
+
+  llvm::StringRef output_filename;
+
+  bool asm_output = false;
+  bool force_obj_output = false;
+  bool custom_core = false;
+  bool dump_shared_values = false;
+  bool dump_tokens = false;
+  bool omit_file_boundary_tokens = false;
+  bool dump_parse_tree = false;
+  bool dump_raw_sem_ir = false;
+  bool dump_sem_ir = false;
+  bool dump_cpp_ast = false;
+  bool dump_llvm_ir = false;
+  bool dump_asm = false;
+  bool dump_mem_usage = false;
+  bool dump_timings = false;
+  bool stream_errors = false;
+  bool preorder_parse_tree = false;
+  bool builtin_sem_ir = false;
+  bool prelude_import = false;
+  bool output_last_input_only = false;
+
+  llvm::SmallVector<llvm::StringRef> exclude_dump_file_prefixes;
+
+  llvm::StringRef sem_ir_crash_dump;
+
+  bool mangle_string_fingerprint = false;
+
+  // Get the LLVM optimization level corresponding to a Carbon optimization
+  // level.
   static auto GetLLVMOptimizationLevel(Lower::OptimizationLevel opt_level)
       -> llvm::OptimizationLevel;
 
   // Get the `-O` flag corresponding to an optimization level.
   static auto GetClangOptimizationFlag(Lower::OptimizationLevel opt_level)
       -> llvm::StringLiteral;
+
+  // Returns a string for printing the phase in a diagnostic.
+  static auto PhaseToString(CompileOptions::Phase phase) -> std::string;
 };
 
 }  // namespace Carbon
 
-#endif  // CARBON_TOOLCHAIN_DRIVER_SHARED_COMPILE_OPTIONS_H_
+#endif  // CARBON_TOOLCHAIN_DRIVER_COMPILE_OPTIONS_H_
